+++ packages/core/src/serializer.ts
@@ -5,13 +5,14 @@ import {
     LionWebJsonMetaPointer,
     LionWebJsonNode
 } from "@lionweb/json"
-import { asArray, keepDefineds } from "@lionweb/ts-utils"
+import { asArray, keepDefineds, lazyMapGet, Nested3Map } from "@lionweb/ts-utils"
 import { asIds } from "./functions.js"
 import { Reader } from "./reading.js"
 import { Node } from "./types.js"
 import { BuiltinPropertyValueSerializer } from "./m3/builtins.js"
-import { allFeaturesOf } from "./m3/functions.js"
+import { inheritsDirectlyFrom } from "./m3/functions.js"
 import {
+    Classifier,
     Containment,
     Enumeration,
     Feature,
@@ -45,6 +46,7 @@ const isPropertyValueSerializer = (value: unknown): value is PropertyValueSerial
  * Type to provide (non-required) options to the serializer.
  */
 export type SerializationOptions = Partial<{
+
     /**
      * Determines whether empty feature values are explicitly serialized or skipped during serialization.
      * (The specification states that empty feature values SHOULD be serialized, but not that they MUST be.)
@@ -62,6 +64,7 @@ export type SerializationOptions = Partial<{
      * Misspelled alias of {@link #propertyValueSerializer}, kept for backward compatibility, and to be deprecated and removed later.
      */
     primitiveTypeSerializer: PropertyValueSerializer
+
 }>
 
 /**
@@ -85,15 +88,33 @@ export const nodeSerializer = <NT extends Node>(reader: Reader<NT>, serializatio
         serializationOptions?.propertyValueSerializer ?? serializationOptions?.primitiveTypeSerializer ?? new BuiltinPropertyValueSerializer()
     const serializeEmptyFeatures = serializationOptions?.serializeEmptyFeatures ?? true
 
+    const languageKey2version2classifierKey2allFeatures: Nested3Map<Feature[]> = {}
+    const memoisedAllFeaturesOf = (classifier: Classifier): Feature[] =>
+        lazyMapGet(
+            lazyMapGet(
+                lazyMapGet(
+                    languageKey2version2classifierKey2allFeatures,
+                    classifier.language.key,
+                    () => ({})
+                ),
+                classifier.language.version,
+                () => ({})
+            ),
+            classifier.key,
+            () => [ ...classifier.features, ...(inheritsDirectlyFrom(classifier).flatMap(memoisedAllFeaturesOf)) ]
+        )
+
     return (nodes: NT[]): LionWebJsonChunk => {
         const serializedNodes: LionWebJsonNode[] = [] // keep nodes as much as possible "in order"
         const ids: { [id: LionWebId]: boolean } = {} // maintain a map to keep track of IDs of nodes that have been serialized
         const languagesUsed: Language[] = []
+        const usedLanguageKey2Version2Boolean: { [key: string]: { [version: string]: boolean } } = {}
         const registerLanguageUsed = (language: Language) => {
-            if (!languagesUsed.some(languageUsed => language.equals(languageUsed))) {
+            const version2Boolean = lazyMapGet<{ [version: string]: boolean }>(usedLanguageKey2Version2Boolean, language.key, () => ({}))
+            if (!version2Boolean[language.version]) {
+                version2Boolean[language.version] = true
                 languagesUsed.push(language)
             }
-            // TODO  could make this more efficient by using a hash table
         }
 
         const visit = (node: NT, parent?: NT) => {
@@ -115,7 +136,7 @@ export const nodeSerializer = <NT extends Node>(reader: Reader<NT>, serializatio
             }
             serializedNodes.push(serializedNode)
             ids[node.id] = true
-            allFeaturesOf(classifier).forEach(feature => {  // TODO  can be made more efficient by caching allFeaturesOf(classifier)
+            memoisedAllFeaturesOf(classifier).forEach(feature => {
                 const value = reader.getFeatureValue(node, feature)
                 const featureLanguage = feature.classifier.language
                 registerLanguageUsed(featureLanguage)