+++ packages/ts-utils/CHANGELOG.md
@@ -4,4 +4,5 @@
 
 * Initial creation and publication of this package, as an extraction and de-duplication from `@lionweb/core`, `@lionweb/utilities`, `@lionweb/class-core`, and `@lionweb/validation`.
 * Add 'mapFrom' function that maps an array to a map, using given key and value functions.
+* Introduce explicit types for `nested{1,2,3}Mapper` functions.